+++ packages/docs/src/stories/templates/dropdown.stories.ts
@@ -5,6 +5,7 @@ export default {
   tags: ['!dev', 'autodocs'],
   title: 'Templates/Dropdown',
   parameters: {
+    relatedLinks: ['components/sd-dropdown', 'components/sd-navigation-item', 'styles/sd-hidden-links'],
     design: {
       type: 'figma',
       url: 'https://www.figma.com/design/YDktJcseQIIQbsuCpoKS4V/Component-Docs?node-id=3000-9863&t=JCsisVFNkWSlhSSN-4'
@@ -14,11 +15,26 @@ export default {
 };
 
 /**
- * ### Dropdown with Navigation Items
  *
  * Example of how to use a dropdown to present a list of navigation options, such as a country selector in the header.
  */
-export const Default = {
+export const DropdownWithNavigationItems = {
+  parameters: {
+    docs: {
+      story: {
+        autoplay: true
+      }
+    }
+  },
+
+  play: ({ canvasElement }: { canvasElement: HTMLElement }) => {
+    customElements.whenDefined('sd-dropdown');
+    const dropdown = canvasElement.querySelector<HTMLElement & { containingElement: HTMLElement }>('#dropdown');
+
+    if (dropdown) {
+      dropdown.containingElement = canvasElement.ownerDocument.body;
+    }
+  },
   render: () => html`
     <style>
       #anchor--templates-dropdown--default .innerZoomElementWrapper {
@@ -56,5 +72,55 @@ export const Default = {
         navigationItems.forEach(item => item.addEventListener('click', handleNavigationItemClick));
       });
     </script>
+  `
+};
+
+/**
+ * Example of how to use a dropdown as a contextual action menu triggered by an icon button. A list of navigation items come with an icon on the left. Use it for making actions like Share, Download, and Delete quickly to scan and access.
+ */
+
+export const MenuWithIcons = {
+  name: 'Dropdown with Menu Items',
+  parameters: {
+    docs: {
+      story: {
+        autoplay: true
+      }
+    }
+  },
+
+  play: ({ canvasElement }: { canvasElement: HTMLElement }) => {
+    customElements.whenDefined('sd-dropdown');
+    const dropdown = canvasElement.querySelector<HTMLElement & { containingElement: HTMLElement }>(
+      '#dropdown-menu-items'
+    );
+
+    if (dropdown) {
+      dropdown.containingElement = canvasElement.ownerDocument.body;
+    }
+  },
+
+  render: () => html`
+    <div class="min-h-[300px]">
+      <sd-dropdown id="dropdown-menu-items" distance="4" rounded open stay-open-on-select>
+        <sd-button variant="secondary" slot="trigger">
+          <sd-icon name="system/more-functions" label="Select the actions"></sd-icon>
+        </sd-button>
+        <sd-menu>
+          <sd-menu-item>
+            <sd-icon name="system/share" slot="icon-indent"></sd-icon>
+            Share
+          </sd-menu-item>
+          <sd-menu-item>
+            <sd-icon name="system/download" slot="icon-indent"></sd-icon>
+            Download
+          </sd-menu-item>
+          <sd-menu-item>
+            <sd-icon name="system/trash" slot="icon-indent"></sd-icon>
+            Delete
+          </sd-menu-item>
+        </sd-menu>
+      </sd-dropdown>
+    </div>
   `
 };