+++ packages/components/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @solid-design-system/components
 
+## 7.4.0
+
 ## 7.3.0
 
 ### Minor Changes